Replace httpretty with responses (#48)
* Replace httpretty with responses

httpretty is not maintained and no wheel is available requiring it be built, replace it with responses.
Remove httpretty and mock.
Use HTTPStatus.value.
Replace remove_None_fields_from_dict, was leaving empty lists and dicts, with remove_empty_fields_from_dict
Remove days option from csr, this option is invalid and not included in the csr.

Signed-off-by: NjM3MjY5NzAgNzA3MzA3 <102704081+NjM3MjY5NzAgNzA3MzA3@users.noreply.github.com>

* Resolve test_event timezone issues

Get the local timezone where test is running.
Add tzdata for Windows.

Signed-off-by: NjM3MjY5NzAgNzA3MzA3 <102704081+NjM3MjY5NzAgNzA3MzA3@users.noreply.github.com>

---------

Signed-off-by: NjM3MjY5NzAgNzA3MzA3 <102704081+NjM3MjY5NzAgNzA3MzA3@users.noreply.github.com>